Add wins cli upgrade

+++ cmd/upgrade/upgrade.go
@@ -17,17 +17,32 @@ $ErrorActionPreference = "Stop";
 $newBinPath = "%[1]s"
 $winsArgs = "%[2]s"
 
-$currBinPath = (Get-CimInstance Win32_Service -Filter 'Name = "rancher-wins"').PathName.Split(' ')[0]
+$winsProcess = (Get-CimInstance Win32_Service -Filter 'Name = "rancher-wins"')
 
-Get-Service -Name rancher-wins | Stop-Service;
-
-sc.exe delete rancher-wins
-
-Copy-Item -Recurse -Force -Path $newBinPath -Destination $currBinPath | Out-Null
+if ($winsProcess -ne $null) {
+	$currBinPath = $winsProcess.PathName.Split(' ')[0]
+	if ($currBinPath -ne $newBinPath) {
+		Get-Service -Name rancher-wins | Stop-Service;
+		sc.exe delete rancher-wins
+		Copy-Item -Recurse -Force -Path $newBinPath -Destination $currBinPath | Out-Null
+	}
+} else {
+	$currBinPath = $newBinPath
+}
 
 "$currBinPath srv app run $winsArgs" | iex
 
 Start-Service -Name rancher-wins;
+}`
+
+const restartServicePS1 = `& {
+	$ErrorActionPreference = "Stop";
+
+	$winsProcess = (Get-CimInstance Win32_Service -Filter 'Name = "rancher-wins"')
+	if ($winsProcess -ne $null) {
+		"$winsProcess" | iex
+		Start-Service -Name rancher-wins;
+	}
 }`
 
 var _upgradeFlags = []cli.Flag{
@@ -40,6 +55,10 @@ var _upgradeFlags = []cli.Flag{
 		Name:  "wins-args",
 		Usage: "[optional] Arguments to pass onto wins srv app run",
 		Value: "--register",
+	},
+	cli.BoolFlag{
+		Name:  "debug",
+		Usage: "[optional] whether to print debugging logs from performing the upgrade",
 	},
 }
 
@@ -66,18 +85,26 @@ func _upgradeRequestParser(cliCtx *cli.Context) (err error) {
 }
 
 func _upgradeAction(cliCtx *cli.Context) (err error) {
+	if cliCtx.Bool("debug") {
+		logrus.SetLevel(logrus.DebugLevel)
+	}
 	binPath := cliCtx.String("binary")
 	winsArgs := cliCtx.String("wins-args")
-	out, err := powershell.RunCommand(fmt.Sprintf(upgradePS1Fmt, binPath, winsArgs))
+	out, err := powershell.RunCommandf(upgradePS1Fmt, binPath, winsArgs)
+	if len(out) > 0 {
+		logrus.Debugf("logs from upgrade.ps1 \n%s\nEOF", out)
+	}
 	if err != nil {
 		logrus.Error("upgrade failed, attempting to ensure rancher-wins service is not stopped: %v", err)
-		_, restartServiceErr := powershell.RunCommand("Start-Service -Name rancher-wins")
+		out, restartServiceErr := powershell.RunCommand(restartServicePS1)
+		if len(out) > 0 {
+			logrus.Debugf("logs from upgrade.ps1 \n%s\nEOF", out)
+		}
 		if err != nil {
 			logrus.Errorf("unable to restart rancher-wins service: %v", restartServiceErr)
 		}
 		return err
 	}
 	logrus.Info("upgrade succeeded")
-	logrus.Info(out)
 	return nil
 }
